mavlink: move SMART_BATTERY_INFO to separate stream header
